Rename classes
https://trello.com/c/JaJkiZQ6/8-time-tracker-ui-create-edit-report-screen
